Several fixes in setup.py and core.py

+++ src/wishbone/core.py
@@ -6,6 +6,7 @@ from numpy import matlib
 
 from sklearn.neighbors import NearestNeighbors
 from scipy import sparse, stats
+from scipy.sparse import csgraph
 from numpy import linalg
 from scipy.sparse.linalg import norm
 from scipy.sparse.csgraph import dijkstra
@@ -226,7 +227,7 @@ def _trajectory_landmarks(spdists, data, s, waypoints, partial_order,
 		if branch:
 			tailk=30
 			tailband = np.where(dijkstra_dist_matrix>=np.percentile(dijkstra_dist_matrix, 85))[0]
-			tailk = min(len(tailband), tailk, np.floor(len(waypoints)/2))
+			tailk = int(min(len(tailband), tailk, np.floor(len(waypoints)/2)))
 			to_replace = np.random.randint(len(waypoints)-1, size=tailk)
 			tailband_sample = np.random.choice( tailband, size=tailk, replace=False)
 			for i in range(len(to_replace)):
@@ -340,7 +341,7 @@ def _splittobranches(trajs, t, data, landmarks, dist, paths_l2l):
 	#return with current branches if branch is not found
 	if(len(c_branch) == 1):
 		print('Branch not found\n')
-		I = np.absolute(dist[:len(landmarks)][:]).min()
+		I = np.absolute(dist[:len(landmarks)][:]).min().astype(np.int)
 		RNK = c[I]
 		Y = np.zeros((len(RNK)))
 		pb = t[landmarks[(np.where(c == c_branch[0])[0][0])]]
